# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  handler.idStruct:
    properties:
      id:
        type: integer
    type: object
  handler.saveActorDTO:
    properties:
      birthDate:
        type: string
      male:
        type: boolean
      name:
        type: string
    type: object
  handler.signInNpit:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.Actor:
    properties:
      birth_date:
        type: string
      films:
        items:
          $ref: '#/definitions/model.Film'
        type: array
      id:
        type: integer
      male:
        type: boolean
      name:
        type: string
    type: object
  model.ActorUpdate:
    properties:
      add_film_ids:
        items:
          type: integer
        type: array
      birth_date:
        type: string
      delete_film_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      male:
        type: boolean
      name:
        type: string
    type: object
  model.Film:
    properties:
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      name:
        maxLength: 150
        minLength: 1
        type: string
      rating:
        maximum: 10
        minimum: 1
        type: integer
      release_date:
        type: string
    type: object
  model.FilmCreate:
    properties:
      actors:
        items:
          type: integer
        type: array
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      name:
        maxLength: 150
        minLength: 1
        type: string
      rating:
        maximum: 10
        minimum: 1
        type: integer
      release_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for films and actors
  title: Filmlib API
  version: "1.0"
paths:
  /actor/all:
    get:
      consumes:
      - application/json
      description: get all actors
      operationId: get-all-actors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Actor'
            type: array
        "500":
          description: Failed to get all actors
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Actors
      tags:
      - Actors
  /actor/delete:
    delete:
      consumes:
      - application/json
      description: Delete an actor by ID
      parameters:
      - description: Actor ID to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.idStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to delete actor
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an actor
      tags:
      - Actors
  /actor/save:
    post:
      consumes:
      - application/json
      description: Save the details of an actor
      parameters:
      - description: Actor object to be saved
        in: body
        name: actor
        required: true
        schema:
          $ref: '#/definitions/handler.saveActorDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the saved actor
          schema:
            $ref: '#/definitions/handler.idStruct'
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to save actor
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Save an actor
      tags:
      - Actors
  /actor/update:
    put:
      consumes:
      - application/json
      description: Update an existing actor
      parameters:
      - description: Actor data to be updated
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ActorUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to update actor
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update actor
      tags:
      - Actors
  /film/all:
    get:
      description: Get all films with the specified sorting parameters
      parameters:
      - description: 'Sort films by: [rating, name, release_date]'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: [asc, desc]'
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/model.Film'
              type: array
            type: array
        "500":
          description: Failed to get films with sort
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all films with sorting
      tags:
      - Film
  /film/delete:
    delete:
      consumes:
      - application/json
      description: Delete a film from the database by its ID
      parameters:
      - description: Film ID to delete
        in: body
        name: filmId
        required: true
        schema:
          $ref: '#/definitions/handler.idStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Film deleted successfully
          schema:
            type: string
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to delete film
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a film by ID
      tags:
      - Film
  /film/name:
    get:
      description: Get films by providing a partial film name and actor name
      parameters:
      - description: Actor name
        in: query
        name: actor
        type: string
      - description: Partial film name
        in: query
        name: film
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of films
          schema:
            items:
              $ref: '#/definitions/model.Film'
            type: array
        "500":
          description: Failed to get films by partial names
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get films by partial name and actor
      tags:
      - Film
  /film/save:
    post:
      consumes:
      - application/json
      description: Save a film to the database
      parameters:
      - description: Film data to save
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/model.FilmCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Film saved successfully
          schema:
            $ref: '#/definitions/handler.idStruct'
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to save film
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Save film
      tags:
      - Film
  /film/update:
    put:
      consumes:
      - application/json
      description: Update a film in the database
      parameters:
      - description: Film object that needs to be updated
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/model.Film'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to update film
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a film
      tags:
      - Film
  /user/auth:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInNpit'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Failed to decode request body. Invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Failed to generate JWT
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignIn
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"